Split utils into dates and dimensions
Create utils/ folder
