+++ index.js
@@ -1,2 +1,2 @@
 export { default } from './src/WeekView/WeekView';
-export { addLocale, createFixedWeekDate } from './src/utils';
+export { addLocale, createFixedWeekDate } from './src/utils/dates';
