+++ src/utils/dimensions.js
@@ -1,13 +1,15 @@
-import moment from 'moment';
-
 export const HEADER_HEIGHT = 50;
 export const CONTENT_OFFSET = 16;
-export const DATE_STR_FORMAT = 'YYYY-MM-DD';
-export const availableNumberOfDays = [1, 3, 5, 7];
 
 const TIMES_WIDTH_PERCENTAGE = 18;
 const PAGE_WIDTH_PERCENTAGE = (100 - TIMES_WIDTH_PERCENTAGE) / 100;
 
+/**
+ *
+ * @param {*} totalWidth
+ * @param {*} numberOfDays
+ * @returns
+ */
 export const computeHorizontalDimensions = (totalWidth, numberOfDays) => {
   // Each day must have an equal width (integer pixels)
   const dayWidth = Math.floor(
@@ -104,51 +106,4 @@ export const yToSeconds = (yValue, verticalResolution, minutesOffset = 0) => {
   const secondsResolution = verticalResolution / 60;
   const secondsInDay = (yValue - CONTENT_OFFSET) / secondsResolution;
   return secondsInDay + minutesOffset * 60;
-};
-
-export const getFormattedDate = (date, format) => {
-  return moment(date).format(format);
-};
-
-export const setLocale = (locale) => {
-  if (locale) {
-    moment.locale(locale);
-  }
-};
-
-export const addLocale = (locale, obj) => {
-  moment.locale(locale, obj);
-};
-
-export const getCurrentMonth = (date) => {
-  return moment(date).format('MMMM Y');
-};
-
-/**
- * Get the amount of minutes in a day of a date.
- * @param {Date} date
- * @returns amount of minutes in the day.
- */
-export const minutesInDay = (date) => {
-  const dateObj = moment(date);
-  if (!dateObj.isValid()) return 0;
-  return dateObj.hours() * 60 + dateObj.minutes();
-};
-
-export const calculateDaysArray = (date, numberOfDays, rightToLeft) => {
-  const dates = [];
-  for (let i = 0; i < numberOfDays; i += 1) {
-    const currentDate = moment(date).add(i, 'd');
-    dates.push(currentDate);
-  }
-  return rightToLeft ? dates.reverse() : dates;
-};
-
-export const createFixedWeekDate = (day, hours, minutes = 0, seconds = 0) => {
-  const date = moment();
-  date.isoWeekday(day);
-  date.hours(hours);
-  date.minutes(minutes);
-  date.seconds(seconds);
-  return date.toDate();
 };
